test: views, models
