+++ discs/templates/discs/artists.html
@@ -6,7 +6,7 @@
 
       <a href="#">+ new artist</a>
       <a href="#">- remove artist</a>
-
+   {% if artists %}
       {% for artist in artists %}
          <p>
             <a href="{% url 'artist-detail' slug=artist.slug %}">
@@ -15,6 +15,9 @@
          </p>
          <span>{{ artist.name }}</span>
       {% endfor %}
+   {% else %}
+         <p>Not artist yet.</p>
+   {% endif %}
 
 
 {% endblock content %}
